ft_printf V1

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: eguelin <eguelin@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/19 11:29:12 by eguelin           #+#    #+#             */
-/*   Updated: 2022/11/20 12:50:44 by eguelin          ###   ########lyon.fr   */
+/*   Updated: 2022/11/20 14:34:17 by eguelin          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define FT_PRINTF_H
 # include <stdio.h>
 # include <stdarg.h>
+# include <unistd.h>
 
 void	ft_putnbr(int n);
 void	ft_putunbr_base(unsigned int unn, unsigned int size_set, char *set);
@@ -24,7 +25,7 @@ int		ft_putnbrlen(int n);
 int		ft_putstr(const char *s);
 int		ft_put_address(size_t n);
 int		ft_printf(const char *format, ...);
-int		ft_unbrlen_base(unsigned int un, unsigned int size_set);
+int		ft_unbrlen_base(size_t un, size_t size_set);
 int		ft_putunbrlen_base(unsigned int n, unsigned int size_set, char *set);
 
 #endif
